move http headers initialisation to it's own package
